Refactor control panel.
