+++ Sources/Nativ/Features/ControlPanel/ControlPanelWindowState.swift
@@ -1,0 +1,61 @@
+import AppKit
+import Combine
+import NativExtensionSDK
+import NativServerKit
+import SwiftUI
+import UniformTypeIdentifiers
+
+struct ControlPanelWindowStateReader: NSViewRepresentable {
+    @Binding var isFullScreen: Bool
+
+    func makeNSView(context: Context) -> ControlPanelWindowStateReaderView {
+        let view = ControlPanelWindowStateReaderView()
+        view.onWindowChange = context.coordinator.update(window:)
+        return view
+    }
+
+    func updateNSView(_ view: ControlPanelWindowStateReaderView, context: Context) {
+        context.coordinator.isFullScreen = $isFullScreen
+        view.onWindowChange = context.coordinator.update(window:)
+        view.reportWindowState()
+    }
+
+    func makeCoordinator() -> Coordinator {
+        Coordinator(isFullScreen: $isFullScreen)
+    }
+
+    @MainActor
+    final class Coordinator {
+        var isFullScreen: Binding<Bool>
+
+        init(isFullScreen: Binding<Bool>) {
+            self.isFullScreen = isFullScreen
+        }
+
+        func update(window: NSWindow?) {
+            window?.titlebarSeparatorStyle = .none
+
+            let newValue = window?.styleMask.contains(.fullScreen) == true
+            guard isFullScreen.wrappedValue != newValue else { return }
+            isFullScreen.wrappedValue = newValue
+        }
+    }
+}
+
+@MainActor
+final class ControlPanelWindowStateReaderView: NSView {
+    var onWindowChange: ((NSWindow?) -> Void)?
+
+    override func viewDidMoveToWindow() {
+        super.viewDidMoveToWindow()
+        reportWindowState()
+
+        DispatchQueue.main.async { [weak self] in
+            self?.reportWindowState()
+        }
+    }
+
+    func reportWindowState() {
+        onWindowChange?(window)
+    }
+}
